% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bummer_stan.R
\name{bummer_stan}
\alias{bummer_stan}
\title{Bayesian BUMMER with Stan}
\usage{
bummer_stan(y, X, y_pred, n_grid = 1000, n_samples = 1000,
  algorithm = "fullrank", pooled = TRUE, vb = TRUE, mcore = TRUE, ...)
}
\arguments{
\item{y}{Numeric matrix of compositional count values.}

\item{X}{Numeric vector of climate values.}

\item{n_samples}{Number of samples}

\item{algorithm}{Variational algorithm of meanfield or fullrank for Stan}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}

\item{output_samples}{Number of samples passed to Stan}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian BUMMER with Stan
}
