% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_batch_diversity}
\alias{compute_batch_diversity}
\title{Function to compute diversity score for each data point}
\usage{
compute_batch_diversity(X, batch_labels, perplexity = 30, ratio_use = 1)
}
\arguments{
\item{X}{data matrix, samples (rows) by features (columns) (e.g. matrix(runif(500), 100, 5))}

\item{batch_labels}{1D array of discrete batch assignment (e.g. c("A", "A", "B", "C", "A", "B"))}

\item{perplexity}{effective number of neighbors (defined in SNE/tSNE algorithms)}

\item{ratio_use}{ratio of data to use. 1 by default. Faster with smaller ratio.}
}
\description{
Function to compute diversity score for each data point
}
