% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2proporptions.R
\name{counts2proportions}
\alias{counts2proportions}
\title{Convert compositional count data to proportions}
\usage{
counts2proportions(counts)
}
\arguments{
\item{counts}{Numeric matrix of integer compositional count values.}
}
\value{
A Numeric matrix of proportional composition values
}
\description{
converts count data to proportions
counts A data frame of pollen counts of size N_sites by N_taxa
returns a data frame of the same size 
Non-zero rows sum to 1
Zero rows remain 0
}
