% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_diff_exp.R
\name{fast_diff_exp}
\alias{fast_diff_exp}
\title{Fast Mann Whitney U Test and auROC}
\usage{
fast_diff_exp(X, y, verbose = TRUE)
}
\arguments{
\item{verbose}{boolean, TRUE for warnings and messages.}

\item{Xt}{sample by feature matrix.}

\item{cols}{vector of group labels.}
}
\value{
table with the following columns: 
\itemize{
\item 
\item 
\item 
}
}
\description{
Compute auROC and p-value based on Gaussian approximation of 
Mann Whitney U statistic
}
\examples{

data(wilcoxauc)
res <- fast_diff_exp(exprs, y)
## Also works on sparse matrices
res <- fast_diff_exp(as(exprs, 'dgCMatrix'), y)
head(res)

}
