% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_wilcox.R
\name{fast_wilcox}
\alias{fast_wilcox}
\title{Fast Mann Whitney U Test and auROC}
\usage{
fast_wilcox(Xt, cols, verbose = TRUE)
}
\arguments{
\item{Xt}{sample by feature matrix.}

\item{cols}{vector of group labels.}

\item{verbose}{boolean, TRUE for warnings and messages.}
}
\value{
list of 3 gene-by-group tables: 
\itemize{
\item pvals: nominal p values
\item fdr: Benjamini-Hochberg corrected p values
\item auc: auROC values
}
}
\description{
Compute auROC and p-value based on Gaussian approximation of 
Mann Whitney U statistic
}
\examples{

data(wilcoxauc)
res <- fast_wilcox(t(exprs), y)
## Also works on sparse matrices
res <- fast_wilcox(as(t(exprs), 'dgCMatrix'), y)
head(res$pvals)
head(res$fdr)
head(res$auc)

}
