% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BUMMER_CV.R
\name{fit_BUMMER_CV}
\alias{fit_BUMMER_CV}
\title{Bayesian BUMMER with Stan}
\usage{
fit_BUMMER_CV(y_train, y_test, X_train, X_test, n_samples = 1000,
  n_grid = 1000, algorithm = "fullrank", pooled = TRUE, vb = TRUE,
  mcore = TRUE)
}
\arguments{
\item{algorithm}{Variational algorithm of meanfield or fullrank for Stan}

\item{y}{Numeric matrix of compositional count values.}

\item{X}{Numeric vector of climate values.}

\item{output_samples}{Number of samples passed to Stan}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian BUMMER with Stan
}
