% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_MAT_CV.R
\name{fit_MAT_CV}
\alias{fit_MAT_CV}
\title{Bayesian BUMMER with Stan}
\usage{
fit_MAT_CV(y_train_prop, y_test_prop, X_train, X_test, k = k, sse = TRUE,
  nboot = nboot, lean = FALSE)
}
\arguments{
\item{y}{Numeric matrix of compositional count values.}

\item{X}{Numeric vector of climate values.}

\item{output_samples}{Number of samples passed to Stan}

\item{algorithm}{Variational algorithm of meanfield or fullrank for Stan}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian BUMMER with Stan
}
