% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-CV.R
\name{make_CV}
\alias{make_CV}
\title{Cross-validation of Compositional datasets}
\usage{
make_CV(y, X, model_name, kfold, n_samples = 1000, n_grid = 1000,
  algorithm = "fullrank", pooled = TRUE, vb = TRUE, mcore = TRUE)
}
\arguments{
\item{y}{Numeric matrix of compositional count values.}

\item{X}{Numeric vector of climate values.}

\item{n_samples}{Number of samples passed to Stan}

\item{n_grid}{Number of grid points for importance sampling prediction}

\item{algorithm}{Variational algorithm of meanfield or fullrank for Stan}

\item{pooled}{Hierarchically pooled version of algorithm}

\item{vb}{Run variational Bayes or NUTS HMC sampler}

\item{mcore}{Enable multicore on a local maching}

\item{y_pred}{Numeric matrix of compositional count values for prediction.}

\item{df}{Integer for the degrees of freedom in the basis expansion}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of
}
\description{
Cross-validation of Compositional datasets
}
