% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCRPSGauss.R
\name{makeCRPSGauss}
\alias{makeCRPSGauss}
\title{Function to evalutate the Continous Ranked Probability Score
for a prediction that has a Gaussian distribution
This function is useful for WA, WAPLS, and MAT 
(and other potential methods that use bootstrapped uncertainty)}
\usage{
makeCRPSGauss(mu, sdev, truth)
}
\arguments{
\item{mu}{Numeric vVctor of predictive means for the hold out data}

\item{sdev}{Numeric Vector of predictive standard deviations for the hold out data}

\item{truth}{Numeric Vector of held out climate variables and is the target of the prediction}
}
\value{
A Numeric Vector of each held-out sample's CRPS score using a Gaussian predictive distribution
}
\description{
Function to evalutate the Continous Ranked Probability Score
for a prediction that has a Gaussian distribution
This function is useful for WA, WAPLS, and MAT 
(and other potential methods that use bootstrapped uncertainty)
}
