\name{methods.package}
\alias{methods.package}
\docType{package}
\title{
Formal Methods and Classes
}
\description{
Formally defined methods and classes for R objects, plus
other programming tools, as described in the reference.
}
\details{
This package provides the so-called \dQuote{S version 4} 
object oriented system.  

For a complete
list of functions and classes, use \code{library(help="methods")}.
}

\references{
``Programming with Data'' (1998), John M. Chambers,
Springer NY. \url{http://www.omegahat.org/RSMethods/Intro.ps} and
\url{http://www.omegahat.org/RSMethods/Intro.pdf}.
}
\author{
R Development Core Team

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
\keyword{ methods }
