% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Muliple ggplot plots in one image}
\usage{
multiplot(..., plotlist = NULL, cols)
}
\arguments{
\item{...}{The set of plots}

\item{plotlist}{A list of plots}

\item{cols}{The number of columns of plots}
}
\value{
A ggplot
}
\description{
Muliple ggplot plots in one image
}
