\name{show}
\alias{show}
\title{Show an Object}
\usage{
show(object)
}
\description{
  display the object, by printing, plotting or whatever suits its class.

This function exists to be specialized by methods; the definition stored here
is intended as the default method.  

  In S-Plus, but not currently in R, \code{show}
is called for automatic display of the result of an evaluated expression.

}
\keyword{programming}
