% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxauc.R
\name{wilcoxauc}
\alias{wilcoxauc}
\title{Fast Mann Whitney U Test and auROC}
\usage{
wilcoxauc(X, ...)
}
\arguments{
\item{Xt}{sample by feature matrix.}

\item{cols}{vector of group labels.}

\item{verbose}{boolean, TRUE for warnings and messages.}
}
\value{
table with the following columns: 
\itemize{
\item 
\item 
\item 
}
}
\description{
Compute auROC and p-value based on Gaussian approximation of 
Mann Whitney U statistic
}
\examples{

data(wilcoxauc)
res <- fast_diff_exp(exprs, y)
## Also works on sparse matrices
res <- fast_diff_exp(as(exprs, 'dgCMatrix'), y)
head(res)

}
