% File src/library/methods/man/Classes.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2016 R Core Team
% Distributed under GPL 2 or later

\name{Classes}
\alias{Classes}
\title{S4 Class Documentation}
\description{
  You have navigated to an old link to documentation of S4 classes.

  For basic use of classes and methods, see \link{Introduction}; to
  create new class definitions, see \code{\link{setClass}}; for
  technical details on S4 classes, see \link{Classes_Details}.

}


\references{
 Chambers, John M. (2016)
 \emph{Extending R},
  Chapman & Hall.
(Chapters 9 and 10.)
}

\keyword{programming}
\keyword{classes}
\keyword{methods}
