% File src/library/methods/man/methods-deprecated.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2009 R Core Team
% Distributed under GPL 2 or later

\name{methods-deprecated}
\alias{methods-deprecated}
%----- NOTE:  ../R/methods-deprecated.R   must be synchronized with this!
\title{Deprecated Functions in Package \pkg{methods}}
%----- PLEASE: put \alias{.} here for EACH ! --- Do not remove this file, even when ``empty''
%
\description{
  These functions are provided for compatibility with older versions of
  \R only, and may be defunct as soon as the next release.
}
\details{
  The original help page for these functions is often
  available at \code{help("oldName-deprecated")} (note the quotes).
  Functions in packages other than the methods package are listed in
  \code{help("pkg-deprecated")}.
}
\seealso{
  \code{\link{Deprecated}}, \code{\link{methods-defunct}}
}
\keyword{internal}
\keyword{misc}
