% File src/library/methods/man/methods-internal.Rd
% Part of the R package, https://www.R-project.org
% Copyright 2026 R Core Team
% Distributed under GPL 2 or later

\name{methods-internal}
\alias{initRefFields}
\title{Internal \pkg{methods} Functions}
\description{
  Internal \pkg{methods} functions.
}
\usage{
initRefFields(.Object, classDef, selfEnv, args)
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
