% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgp_stan.R
\name{mvgp_stan}
\alias{mvgp_stan}
\title{Bayesian MVGP with Stan}
\usage{
mvgp_stan(y, X, y_pred, df = 6, n_grid = 1000, n_samples = 1000,
  algorithm = "fullrank", pooled = TRUE, vb = TRUE, mcore = TRUE, ...)
}
\arguments{
\item{y}{Numeric matrix of compositional count values.}

\item{X}{Numeric vector of climate values.}

\item{y_pred}{Numeric matrix of compositional count values for prediction.}

\item{df}{Integer for the degrees of freedom in the basis expansion}

\item{n_grid}{Number of grid locations for importance sampling}

\item{n_samples}{Number of samples passed to Stan}

\item{algorithm}{Variational algorithm of meanfield or fullrank for Stan}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian MVGP with Stan
}
