% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxauc.R
\name{wilcoxauc}
\alias{wilcoxauc}
\title{Fast Wilcoxon rank sum test and auROC}
\usage{
wilcoxauc(X, ...)
}
\arguments{
\item{X}{A feature-by-sample matrix, Seurat object, or SingleCellExperiment object}

\item{y}{vector of group labels.}

\item{groups_use}{(optional) which groups from y vector to test.}

\item{group_by}{(Seurat & SCE) name of groups variable ('e.g. Cluster').}

\item{assay}{(Seurat & SCE) name of feature matrix slot (e.g. 'data' or 'logcounts').}

\item{seurat_assay}{(Seurat) name of Seurat Assay (e.g. 'RNA').}

\item{verbose}{boolean, TRUE for warnings and messages.}
}
\value{
table with the following columns: 
\itemize{
\item 
\item 
\item 
}
}
\description{
Computes auROC and Wilcoxon p-value based on Gaussian approximation
}
\examples{

library(presto)
data(exprs)
data(y)

}
